cogs/anime_quote.py
import discord
from discord.ext import commands

class quotes(commands.Cog):

    def __init__(self, client):
        self.client = client

    @commands.Cog.listener()
    async def on_ready(self):
        print(f"Anime cogs loaded successfully!")

    @commands.command()
    async def ping(self, ctx):
        await ctx.send('pong')

def setup(bot):
    bot.add_cog(quotes(bot))
